requirements.txt
python-whois
dnspython
wafw00f
wappalyzer
pandas
